Move live server test case to its own module.
